rename the component to .jsx file extension
