+++ package.json
@@ -102,6 +102,10 @@
     "css-loader": "^0.28.4",
     "enzyme": "^2.9.1",
     "eslint": "^4.1.1",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.3.0",
     "expect": "^1.20.2",
     "extract-text-webpack-plugin": "^3.0.0",
     "gulp": "^3.9.1",
